make fundAccounts TS and use @emizzle's async function
