+++ packages/embark-accounts-manager/src/fundAccount.ts
@@ -1,0 +1,31 @@
+import Web3 = require("web3");
+
+const TARGET = 0x7FFFFFFFFFFFFFFF;
+
+export default async function fundAccount(web3: Web3, accountAddress: string, coinbaseAddress: string, hexBalance: string | number | undefined) {
+  if (!hexBalance) {
+    hexBalance = TARGET;
+  }
+  const targetBalance = web3.utils.toBN(hexBalance);
+
+  // check if account is already funded
+  let accountBalance = await web3.eth.getBalance(accountAddress);
+  accountBalance = web3.utils.toBN(accountBalance);
+  if (accountBalance.gte(targetBalance)) {
+    return;
+  }
+
+  // run in parallel
+  const getGasPricePromise = web3.eth.getGasPrice();
+  const getTxCountPromise = web3.eth.getTransactionCount(coinbaseAddress);
+
+  const [gasPrice, lastNonce] = await Promise.all([getGasPricePromise, getTxCountPromise]);
+
+  return web3.eth.sendTransaction({
+    from: coinbaseAddress,
+    to: accountAddress,
+    value: targetBalance.sub(accountBalance),
+    gasPrice: gasPrice,
+    nonce: lastNonce
+  });
+}
